+++ packages/context/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @doerjs/context
 
+## 0.1.1
+
+### Patch Changes
+
+- @doerjs/cli@0.1.17
+
 ## 0.1.0
 
 ### Minor Changes