+++ packages/router-webpack-plugin/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @doerjs/router-webpack-plugin
 
+## 0.1.1
+
+### Patch Changes
+
+- 调整路由地址映射规则为按目录进行映射，同时取消路由规则参数
+
 ## 0.1.0
 
 ### Minor Changes